moved ffilelength from d_io.cpp to cfile.cpp and removed d_io.cpp from the project
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@13240 01821946-851b-0410-b7f4-f1477c70e57a

+++ input/include/event.h
@@ -20,7 +20,7 @@
 #endif
 
 int event_init();
-void event_poll(unsigned long mask);
+void event_poll(uint mask);
 
 #ifdef GII_XWIN
 void init_gii_xwin(Display *disp,Window win);
